记录程序的运行时间
任意编写一段程序，要求记录并输出该段程序执行的时间。
(float)(end - start) 这里要做强制类型转换，否则会有问题。。

+++ chap4/chap4-1.12.cpp
@@ -25,7 +25,7 @@ int main()
 
 	end = clock();//标记时间戳
 
-	printf("The time was: end:%d,,start:%d,,CLK_TCK:%d, ,,,%0.3f \r\n",end,start,CLK_TCK, ((end - start) / CLK_TCK));
+	printf("The time was: end:%d,,start:%d,,CLK_TCK:%d, ,,,%0.3f \r\n",end,start,CLK_TCK, ((float)(end - start) / CLK_TCK));
 
 	float ft = (end - start) / CLK_TCK;
 	cout << "time was:" << ft << endl;
